LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entries-after-srcdoc-navigation-expected.txt



FAIL entries() after setting a srcdoc attribute assert_equals: expected 2 but got 0
